This is a large, dirty commit. The purpose of this commit is to deal with an issue found when analyzing the 3D data which had a problem with extending the whisker during contact point finding. That issue was a problem with BackProject and the order of its arguments. Back project should now assum front is left is A camera.
This allowed me to use splinefit smoothing to the whisker which no longer causes weird shapes. Extension is automatically off during splinefit.

Extension during CP finding is now a quadratic fit to the tip portion of the whisker, and only the extended points are changed, not the whole whisker.

interp3Dwhisker was modified to use interparc, but it should still not be used.

added a function to remove the second point in the 3D whisker since it is very often a kink.

added a script verifyMergeQuality to use images and backproject to see how well th 3D merged whisker matches with the images.

since framesize will no longer be only 640x480, edited whisk2merge to keep the frame_size with the data. This is crucial for manipulator finding. Get 3DCP now needs the frame size

Made a utility function to ensure that a whisker struc has column vectors.

edited merge2E3d to account for all these changes, and to no longer do contact finding in 3D. Rather that will happen in 2D.